Fix also other function
